requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coverage==6.3.2
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
django-environ==0.8.1
django-extensions==3.1.5
django-localflavor==3.1
djangorestframework==3.13.1
docker==5.0.3
idna==3.3
oauthlib==3.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.8
websocket-client==1.3.1
Werkzeug==2.0.3
